msglist [nfc]: Move MessageListPage to widgets/message_list from widgets/app
