+++ lib/widgets/message_list.dart
@@ -8,9 +8,45 @@ import '../model/message_list.dart';
 import '../model/narrow.dart';
 import '../model/store.dart';
 import 'action_sheet.dart';
+import 'compose_box.dart';
 import 'content.dart';
+import 'page.dart';
 import 'sticky_header.dart';
 import 'store.dart';
+
+class MessageListPage extends StatelessWidget {
+  const MessageListPage({super.key, required this.narrow});
+
+  static Route<void> buildRoute({required BuildContext context, required Narrow narrow}) {
+    return MaterialAccountPageRoute(context: context,
+      builder: (context) => MessageListPage(narrow: narrow));
+  }
+
+  final Narrow narrow;
+
+  @override
+  Widget build(BuildContext context) {
+    return Scaffold(
+      appBar: AppBar(title: const Text("All messages")),
+      body: Builder(
+        builder: (BuildContext context) => Center(
+          child: Column(children: [
+            MediaQuery.removePadding(
+              // Scaffold knows about the app bar, and so has run this
+              // BuildContext, which is under `body`, through
+              // MediaQuery.removePadding with `removeTop: true`.
+              context: context,
+
+              // The compose box pads the bottom inset.
+              removeBottom: true,
+
+              child: Expanded(
+                child: MessageList(narrow: narrow))),
+
+            ComposeBox(narrow: narrow),
+          ]))));
+  }
+}
 
 class MessageList extends StatefulWidget {
   const MessageList({super.key, required this.narrow});
